pyproject.toml
[project]
name = "thecargo"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyjwt[crypto]>=2.10.0",
    "httpx>=0.28.0",
    "redis>=5.0.0",
    "aio-pika>=9.5.0",
    "minio>=7.2.0",
    "python-multipart>=0.0.12",
]

[tool.hatch.build.targets.wheel]
packages = ["thecargo"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
